Make DynatraceApiRequestThreshold uint16 So the type check is done by the API not us (negative values make no sense) It is only uint16 so the conversion is easier (less to overflow) The max uint16 is 65535, which in this context would be ~45 days, which IMO is fine

+++ pkg/api/v1beta1/dynakube/convert_from_test.go
@@ -270,7 +270,7 @@ func getNewDynakubeBase() dynakube.DynaKube {
 			},
 			TrustedCAs:                   "trusted-ca",
 			NetworkZone:                  "network-zone",
-			DynatraceApiRequestThreshold: address.Of(42),
+			DynatraceApiRequestThreshold: address.Of(uint16(42)),
 			MetadataEnrichment: dynakube.MetadataEnrichment{
 				Enabled:           address.Of(true),
 				NamespaceSelector: getTestNamespaceSelector(),
