+++ pkg/api/v1beta3/dynakube/dynakube_types.go
@@ -80,7 +80,7 @@ type DynaKubeSpec struct { //nolint:revive
 	// Configuration for thresholding Dynatrace API requests.
 	// +kubebuilder:validation:Optional
 	// +operator-sdk:csv:customresourcedefinitions:type=spec,displayName="Dynatrace API Request Threshold",order=9,xDescriptors={"urn:alm:descriptor:com.tectonic.ui:advanced"}
-	DynatraceApiRequestThreshold *int `json:"dynatraceApiRequestThreshold,omitempty"`
+	DynatraceApiRequestThreshold *uint16 `json:"dynatraceApiRequestThreshold,omitempty"`
 
 	// When an (empty) ExtensionsSpec is provided, the extensions related components (extensions controller and extensions collector)
 	// are deployed by the operator.
